move additionalProps to util
